Code refactor of upcoming posts
