+++ src/components/PostCard.tsx
@@ -25,23 +25,15 @@ export const DraftPostMark = () => (
   </div>
 );
 
-export const FuturePostMark = ({ date }: FuturePostProps) => {
-  const postDate = new Date(date);
-
-  const postDay = postDate.getDate();
-
-  const postDateString = postDate.toLocaleDateString("en-GB");
-
-  return (
-    <div className="future-post">
-      <div className="triangle triangle-future  triangle-item">
-        <span className="triangle-text-future triangle-text">
-          {postDateString}
-        </span>
-      </div>
+export const FuturePostMark = ({ date }: FuturePostProps) => (
+  <div className="future-post">
+    <div className="triangle triangle-future  triangle-item">
+      <span className="triangle-text-future triangle-text">
+        {new Date(date).toLocaleDateString("en-GB")}
+      </span>
     </div>
-  );
-};
+  </div>
+);
 
 const shimmer = (width: number, height: number) => `
   <svg width="${width}" height="${height}" version="1.1" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
